fix: Segfaults due to bad refs to vector elements in DummyGraph
The DUmmyGraph constructor builds the blocks vector from size 0 to 5. In
some compilers, this may kick the resizing login which may invalidate
some push_successor refs (that's what was happening).

Either the vector shouldn't resize (possible only for default-insertible
elems -- not DummyBlock)
Or the refs should be used when the size is fixed.

Partially helps with #1